Fix a bug in the Base32 encoder that can cause improper encoding of the first byte (flutter/engine#28328)

+++ engine/src/flutter/fml/base32.cc
@@ -25,7 +25,7 @@ std::pair<bool, std::string> Base32Encode(std::string_view input) {
   output.reserve(encoded_length);
 
   Base32EncodeConverter converter;
-  converter.Append(input[0]);
+  converter.Append(static_cast<uint8_t>(input[0]));
   size_t next_byte_index = 1;
 
   while (converter.CanExtract()) {
